comm-coach/requirements.txt
streamlit>=1.30
streamlit-audiorecorder>=0.0.6
numpy
pandas
matplotlib
python-dotenv
pydub
faster-whisper
ctranslate2
openai